+++ src/helpers/convert-pcre.ts
@@ -4,11 +4,12 @@ export default function pcre(regex: string): RegExp {
 	const replace = (search: string | RegExp, replace: string) => finalRegex = finalRegex.replace(search, replace);
 	const finalFlags = new Set<string>();
 	// Convert inline flag declarations
-	const inlineMatches = regex.matchAll(/\?([a-z]):/g);
-	const startMatches = regex.matchAll(/\(\?([a-z]+)\)/g);
-	for (const [match, flags] of [...inlineMatches, ...startMatches]) {
+	const inlineMatches = regex.matchAll(/\?(-)?([a-z]):/g);
+	const startMatches = regex.matchAll(/\(\?(-)?([a-z]+)\)/g);
+	for (const [match, isNegative, flags] of [...inlineMatches, ...startMatches]) {
 		replace(match, '');
-		[...flags].forEach(flag => finalFlags.add(flag));
+		const func = (flag: string) => isNegative ? finalFlags.delete(flag) : finalFlags.add(flag);
+		[...flags].forEach(func);
 	}
 	// Remove PCRE-only syntax
 	replace(/([*+]){2}/g, '$1');